backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
celery>=5.3.0
redis>=4.5.0
elasticsearch>=8.8.0
yt-dlp>=2023.11.0
Pillow>=10.0.0
python-dateutil>=2.8.2
pytz>=2023.3
drf-spectacular>=0.26.0
django-celery-beat>=2.5.0
requests>=2.31.0
pyotp>=2.9.0
qrcode>=7.4.0
reportlab>=4.0.0
mutagen>=1.47.0
pylast>=5.2.0
psutil>=5.9.0
whitenoise>=6.5.0
